Remove methods on interface structs only used for tests
