+++ chains/solana/contracts/programs/ccip-router/src/instructions/v1/messages.rs
@@ -150,6 +150,18 @@ pub mod ramps {
             }
         }
 
+        fn as_u8_28(single: U256) -> [u8; 28] {
+            single.to_be_bytes()[4..32].try_into().unwrap()
+        }
+
+        impl TimestampedPackedU224 {
+            pub fn from_single(timestamp: i64, single: U256) -> Self {
+                let mut value = [0u8; 28];
+                value.clone_from_slice(&single.to_be_bytes()[4..32]);
+                Self { value, timestamp }
+            }
+        }
+
         #[test]
         fn message_not_validated_for_disabled_destination_chain() {
             let mut chain = sample_dest_chain();
@@ -255,10 +267,10 @@ pub mod ramps {
             // they were used to retrieve correctly dimensioned values)
 
             let arbitrary_timestamp = 100;
-            let usd_per_token = crate::TimestampedPackedU224::from_single(
-                arbitrary_timestamp,
-                U256::new(19816680000000000000),
-            );
+            let usd_per_token = TimestampedPackedU224 {
+                timestamp: arbitrary_timestamp,
+                value: as_u8_28(U256::new(19816680000000000000)),
+            };
 
             BillingTokenConfig {
                 enabled: true,
